+++ src/backend/core/warren/tests/conftest.py
@@ -4,4 +4,3 @@
 
 from .fixtures.app import http_client  # noqa: F401
 from .fixtures.asynchronous import anyio_backend  # noqa: F401
-from .fixtures.backends import es_client, no_cached_es_client  # noqa: F401