+++ src/backend/plugins/video/tests/test_api.py
@@ -1,11 +1,17 @@
 """Tests for the video API endpoints."""
+import json
+from typing import List
+from urllib.parse import ParseResult, quote_plus, unquote, urlencode, urlparse
 
 import arrow
 import pytest
+from httpx import AsyncClient
+from pytest_httpx import HTTPXMock
+from ralph.models.xapi import VideoPlayed
+from ralph.models.xapi.concepts.verbs.video import PlayedVerb
 from warren_video.api import VideoDayViews, VideoViews
-from warren_video.factories import VideoPlayedFactory
 
-from warren.conf import settings
+from warren.backends import lrs_client
 from warren.filters import DatetimeRange
 
 
@@ -20,9 +26,18 @@ from warren.filters import DatetimeRange
         ("%2Ffoo%2Fbar", "/foo/bar"),
     ],
 )
-async def test_views_invalid_video_id_path(video_id, parsed_video_id, http_client):
+async def test_views_invalid_video_id_path(
+    http_client: AsyncClient, video_id: str, parsed_video_id: str
+):
     """Test the video views endpoint with an invalid "video_id" path."""
-    response = await http_client.get(f"/api/v1/video/{video_id}/views")
+    date_range = DatetimeRange()
+    params = {
+        "since": date_range.since.astimezone().isoformat(),
+        "until": date_range.until.astimezone().isoformat(),
+    }
+
+    response = await http_client.get(f"/api/v1/video/{video_id}/views", params=params)
+
     assert response.status_code == 422
     assert (
         response.json().get("detail")[0].get("msg")
@@ -30,7 +45,6 @@ async def test_views_invalid_video_id_path(video_id, parsed_video_id, http_clien
     )
 
 
-# pylint: disable=unused-argument
 @pytest.mark.anyio
 @pytest.mark.parametrize(
     "video_id",
@@ -42,97 +56,104 @@ async def test_views_invalid_video_id_path(video_id, parsed_video_id, http_clien
     ],
 )
 async def test_views_valid_video_id_path_but_no_matching_video(
-    video_id, es_client, http_client
+    http_client: AsyncClient, httpx_mock: HTTPXMock, video_id
 ):
     """Test the video views endpoint with a valid "video_id" but no results."""
 
-    default_range = DatetimeRange()
-    no_statements_response = VideoViews(
+    date_range = DatetimeRange()
+
+    base_url = "http://fake-lrs.com"
+    lrs_client.base_url = base_url
+
+    params = {
+        "since": date_range.since.astimezone().isoformat(),
+        "until": date_range.until.astimezone().isoformat(),
+    }
+
+    query_string = (
+        f"verb={quote_plus(PlayedVerb().id, safe='/')}&activity="
+        f"{quote_plus(unquote(video_id), safe='/')}"
+    )
+
+    lrs_url = ParseResult(
+        scheme=urlparse(lrs_client.base_url).scheme,
+        netloc=urlparse(lrs_client.base_url).netloc,
+        path=lrs_client.statements_endpoint,
+        query=f"{query_string}&{urlencode({**params, 'limit': 500})}",
+        params="",
+        fragment="",
+    ).geturl()
+
+    httpx_mock.add_response(
+        url=lrs_url, method="GET", json={"statements": []}, status_code=200
+    )
+
+    response = await http_client.get(f"/api/v1/video/{video_id}/views", params=params)
+    no_statement_response = VideoViews(
         total=0,
         daily_views=[
             VideoDayViews(day=day.format("YYYY-MM-DD"))
-            for day in arrow.Arrow.range(
-                "day", default_range.since, default_range.until
-            )
+            for day in arrow.Arrow.range("day", date_range.since, date_range.until)
         ],
     )
 
-    response = await http_client.get(f"/api/v1/video/{video_id}/views")
     assert response.status_code == 200
-    assert VideoViews.parse_obj(response.json()) == no_statements_response
+    assert VideoViews.parse_obj(response.json()) == no_statement_response
 
 
 @pytest.mark.anyio
-async def test_views_backend_query(es_client, http_client):
+async def test_views_backend_query(
+    http_client: AsyncClient,
+    httpx_mock: HTTPXMock,
+    video_statements: List[VideoPlayed],
+):
     """Test the video views endpoint backend query results."""
 
-    await VideoPlayedFactory.save_many(
-        es_client,
-        [
-            VideoPlayedFactory.build(
-                [
-                    {"object": {"id": "uuid://ba4252ce-d042-43b0-92e8-f033f45612ee"}},
-                    {
-                        "result": {
-                            "extensions": {
-                                "https://w3id.org/xapi/video/extensions/time": 0.3
-                            }
-                        }
-                    },
-                    {"timestamp": "2023-02-03T15:34:00.000-01:00"},
-                ]
-            ),
-            VideoPlayedFactory.build(
-                [
-                    {"object": {"id": "uuid://ba4252ce-d042-43b0-92e8-f033f45612ee"}},
-                    {
-                        "result": {
-                            "extensions": {
-                                "https://w3id.org/xapi/video/extensions/time": 12.4
-                            }
-                        }
-                    },
-                    {"timestamp": "2023-02-07T19:34:00.000-01:00"},
-                ]
-            ),
-            VideoPlayedFactory.build(
-                [
-                    {"object": {"id": "uuid://ba4252ce-d042-43b0-92e8-f033f45612ee"}},
-                    {
-                        "result": {
-                            "extensions": {
-                                "https://w3id.org/xapi/video/extensions/time": 33.1
-                            }
-                        }
-                    },
-                    {"timestamp": "2023-02-07T15:28:00.000-01:00"},
-                ]
-            ),
-            VideoPlayedFactory.build(
-                [
-                    {"object": {"id": "uuid://8efd1802-dce0-4e81-88e4-39760e683fd8"}},
-                ]
-            ),
+    video_id = "uuid://ba4252ce-d042-43b0-92e8-f033f45612ee"
+
+    date_range = DatetimeRange()
+
+    base_url = "http://fake-lrs.com"
+    lrs_client.base_url = base_url
+
+    params = {
+        "since": date_range.since.astimezone().isoformat(),
+        "until": date_range.until.astimezone().isoformat(),
+    }
+
+    query_string = (
+        f"verb={quote_plus(PlayedVerb().id, safe='/')}&activity="
+        f"{quote_plus(unquote(video_id), safe='/')}"
+    )
+
+    lrs_url = ParseResult(
+        scheme=urlparse(lrs_client.base_url).scheme,
+        netloc=urlparse(lrs_client.base_url).netloc,
+        path=lrs_client.statements_endpoint,
+        query=f"{query_string}&{urlencode({**params, 'limit': 500})}",
+        params="",
+        fragment="",
+    ).geturl()
+
+    statements = [json.loads(statement.json()) for statement in video_statements]
+    httpx_mock.add_response(
+        url=lrs_url, method="GET", json={"statements": statements}, status_code=200
+    )
+
+    response = await http_client.get(f"/api/v1/video/{video_id}/views", params=params)
+
+    assert response.status_code == 200
+
+    video_views = VideoViews.parse_obj(response.json())
+
+    expected_video_views = VideoViews(
+        total=0,
+        daily_views=[
+            VideoDayViews(day=day.format("YYYY-MM-DD"))
+            for day in arrow.Arrow.range("day", date_range.since, date_range.until)
         ],
     )
-    await es_client.indices.refresh(index=settings.ES_INDEX)
+    expected_video_views.daily_views[-1].views = 2
+    expected_video_views.total = 2
 
-    response = await http_client.get(
-        "/api/v1/video/uuid://ba4252ce-d042-43b0-92e8-f033f45612ee/views",
-        params={"since": "2023-02-01", "until": "2023-02-10"},
-    )
-    assert response.status_code == 200
-    video_views = VideoViews.parse_obj(response.json())
-    assert video_views.total == 2
-    assert video_views.daily_views == [
-        VideoDayViews(day="2023-02-01", views=0),
-        VideoDayViews(day="2023-02-02", views=0),
-        VideoDayViews(day="2023-02-03", views=1),
-        VideoDayViews(day="2023-02-04", views=0),
-        VideoDayViews(day="2023-02-05", views=0),
-        VideoDayViews(day="2023-02-06", views=0),
-        VideoDayViews(day="2023-02-07", views=1),
-        VideoDayViews(day="2023-02-08", views=0),
-        VideoDayViews(day="2023-02-09", views=0),
-        VideoDayViews(day="2023-02-10", views=0),
-    ]
+    assert video_views == expected_video_views