+++ src/backend/plugins/video/warren_video/__init__.py
@@ -1,3 +1,3 @@
-"""Warren Video plugin."""
+"""Warren Video plugin main package."""
 
 __version__ = "0.1.0"